requirements.txt
RPi.GPIO==0.7.1
Flask==2.0.1
requests==2.26.0
numpy==1.21.0
pynmea2==1.18.0
Werkzeug==2.0.1
Jinja2==3.0.1
itsdangerous==2.0.1
click==8.0.1
MarkupSafe==2.0.1
opencv-python-headless==4.5.3.56
pytest==6.2.5